refactor: Move research model and schema to the research folder
