+++ research/schema/__init__.py
@@ -1,0 +1,1 @@
+"""Research schema package."""
